be/pyproject.toml
[project]
name = "be"
version = "0.1.0"
description = "🦤 Dodo Bird Travel Planner"
authors = [
    { name = "hongye.an", email = "annhuny@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "nvidia-nat[all]>=1.2.1",
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
    "websockets>=11.0.0",
    "aiofiles>=23.0.0",
    "typing-extensions>=4.8.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.3.27",
    "langchain-openai>=0.2.0",
    "requests>=2.31.0",
    "json-repair>=0.50.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
